API endpoint verified to be working✅

+++ customer.py
@@ -7,7 +7,7 @@ class Customer(BaseModel):
     age: int
     tenure: int
     balance: float
-    products_number: float
+    products_number: int
     credit_card: int
     active_member: int
     estimated_salary: float
